add: move option for player
