+++ src/index.js
@@ -1,6 +1,6 @@
 import './styles/style.scss';
 // import gsap from 'gsap';
-import Controller from './controller';
+import Controller from './js/controller';
 
 const controller = new Controller();
 controller.startBtnEventListener();
