--- src/player.js
@@ -1,18 +1,0 @@
-const canvas = document.querySelector('canvas');
-const ctx = canvas.getContext('2d');
-
-export default class Player {
-  constructor(x, y, radius, color) {
-    this.x = x;
-    this.y = y;
-    this.radius = radius;
-    this.color = color;
-  }
-
-  draw() {
-    ctx.beginPath();
-    ctx.arc(this.x, this.y, this.radius, 0, Math.PI * 2, false);
-    ctx.fillStyle = this.color;
-    ctx.fill();
-  }
-}
