+++ src/styles/style.scss
@@ -1,3 +1,6 @@
+@import "~normalize.css";
+@import "roboto.css";
+
 * {
   margin: 0;
   padding: 0;
